+++ js/AppView.js
@@ -1,30 +1,23 @@
-define(["AppUtils", "AppModel", "RouterManager"],
-	function (AppUtils, AppModel, RouterManager) {
+define(["AppUtils", "AppModel", "Shell", "RouterManager"],
+	function (AppUtils, AppModel, Shell, RouterManager) {
 		const conf = AppUtils.getComponentConfiguration("app", {
 			selector: "body"
 		});
 		return ng.core.Component(conf).Class({
-			constructor: [AppModel, RouterManager, ng.http.Http,
-				function AppView(AppModel, RouterManager, Http) {
+			constructor: [AppModel, Shell, RouterManager,
+				function AppView(AppModel, Shell, RouterManager) {
 					this.model = AppModel;
+					this._shell = Shell;
 					this._routerManager = RouterManager;
-					this._http = Http;
-					this._navigationEndEventEmitter = null;
-					this.currentView = "list";
+					this.currentView = null;
 				}
 			],
 			ngOnInit: function () {
-				const pkg = AppUtils.getPackageFile();
+				const cc = console;
+				cc.dir(process.argv);
+				this._shell.init();
 				this._navigationEndEventEmitter = this._routerManager.on("NAVIGATION_END").subscribe((event) => {
 					this.currentView = event.toUrl.substring(1).split("/")[0];
-				});
-				this._http.get(AppUtils.getRemotePackageUrl()).subscribe((result) => {
-					const remotePkg = result.json();
-					if (remotePkg.version > pkg.version) {
-						if (confirm("New version " + remotePkg.version + " available, do you want download it ?") === true) {
-							nw.Shell.openExternal(AppUtils.getRemoteApplicationDownloadUrl(remotePkg.version));
-						}
-					}
 				});
 				if (this.model.config.games.length === 0) {
 					this.showView("config");